fix for issue-1865 - dereferencing files with dots
