+++ modules/swagger-parser-v3/src/test/resources/issue-1621/example.openapi.yaml
