+++ modules/swagger-parser-v3/src/test/resources/issue-1865/apis/foo.yaml
@@ -1,0 +1,24 @@
+#broken-ref/apis/foo.yaml
+openapi: 3.0.0
+info:
+  title: swagger parser bug, worked until 2.0.28 but not since 2.0.29
+  version: '1.0'
+
+paths:
+
+  '/foo':
+    post:
+      summary: create foo
+      description: Create a new foo.
+      responses:
+        '200':
+          description: a Foo
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/Foo'
+
+components:
+  schemas:
+    Foo:
+      $ref: ../models/foo.model.yaml
