+++ modules/swagger-parser-v3/src/test/resources/issue-1865/examples/foo.example.yaml
@@ -1,0 +1,3 @@
+# broken-ref/examples/foo.example.yaml
+value:
+  foo: bar
