+++ modules/swagger-parser-v3/src/test/resources/issue-1865/models/foo.model.yaml
@@ -1,0 +1,6 @@
+#broken-ref/models/foo.model.yaml
+title: Foo
+type: object
+properties:
+  foo:
+    type: string
