+++ modules/swagger-parser-v3/src/test/resources/issue-1865/openapi30.yaml
@@ -1,0 +1,9 @@
+# broken-ref/openapi30.yaml
+openapi: 3.0.3
+info:
+  title: ref chain
+  version: '1'
+
+paths:
+  /foo:
+    $ref: ./apis/foo.yaml#/paths/~1foo
